feat: added average and progress calculation logic

+++ src/main/java/br/com/unex/edutrack/dto/subject/SubjectResponseDto.java
@@ -1,6 +1,8 @@
 package br.com.unex.edutrack.dto.subject;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
+
 
 public record SubjectResponseDto(
         int id,
@@ -8,7 +10,9 @@ public record SubjectResponseDto(
         int progress,
         LocalDateTime createdAt,
         LocalDateTime updatedAt,
-        float average
+        BigDecimal average
 
 ) {
+
+
 }
